close #57 SpeedCalculatorのpwm算出を左右の車輪ごとに行うように変更

+++ test/ColorStraightTest.cpp
@@ -1,7 +1,7 @@
 /**
  *  @file   ColorStraightTest.cpp
  *  @brief  ColorStraightクラスのテスト
- *  @author Negimarge
+ *  @author Negimarge bizyutyu
  */
 
 #include <gtest/gtest.h>
@@ -23,7 +23,7 @@ namespace etrobocon2023_test {
     // 初期値から期待する走行距離を求める
     int initialRightCount = Measurer::getRightCount();
     int initialLeftCount = Measurer::getLeftCount();
-    int expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
+    double expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
 
     srand(9037);  // 最初に緑を取得する乱数シード
     cs.run();     // 緑まで直進を実行
@@ -31,7 +31,7 @@ namespace etrobocon2023_test {
     // 直進後の走行距離
     int rightCount = Measurer::getRightCount();
     int leftCount = Measurer::getLeftCount();
-    int actual = Mileage::calculateMileage(rightCount, leftCount);
+    double actual = Mileage::calculateMileage(rightCount, leftCount);
 
     EXPECT_EQ(expected, actual);  // 直進前後で走行距離に変化はない
   }
@@ -46,7 +46,7 @@ namespace etrobocon2023_test {
     // 初期値から期待する走行距離を求める
     int initialRightCount = Measurer::getRightCount();
     int initialLeftCount = Measurer::getLeftCount();
-    int expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
+    double expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
 
     /**
      * 最初10回の色取得分の走行距離を許容誤差とする
@@ -61,7 +61,7 @@ namespace etrobocon2023_test {
     // 直進後の走行距離
     int rightCount = Measurer::getRightCount();
     int leftCount = Measurer::getLeftCount();
-    int actual = Mileage::calculateMileage(rightCount, leftCount);
+    double actual = Mileage::calculateMileage(rightCount, leftCount);
 
     EXPECT_LT(expected, actual);          // 実行後に少しでも進んでいる
     EXPECT_GE(expected + error, actual);  // 直進後の走行距離が許容誤差以内である
@@ -77,7 +77,7 @@ namespace etrobocon2023_test {
     // 初期値から期待する走行距離を求める
     int initialRightCount = Measurer::getRightCount();
     int initialLeftCount = Measurer::getLeftCount();
-    int expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
+    double expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
 
     /**
      * 最初10回の色取得分の走行距離を許容誤差とする
@@ -92,7 +92,7 @@ namespace etrobocon2023_test {
     // 直進後の走行距離
     int rightCount = Measurer::getRightCount();
     int leftCount = Measurer::getLeftCount();
-    int actual = Mileage::calculateMileage(rightCount, leftCount);
+    double actual = Mileage::calculateMileage(rightCount, leftCount);
 
     EXPECT_GT(expected, actual);          // 実行後に少しでも進んでいる
     EXPECT_LE(expected + error, actual);  // 直進後の走行距離が許容誤差以内である
@@ -107,7 +107,7 @@ namespace etrobocon2023_test {
     // 初期値から期待する走行距離を求める
     int initialRightCount = Measurer::getRightCount();
     int initialLeftCount = Measurer::getLeftCount();
-    int expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
+    double expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
 
     // Warning文
     string expectedOutput = "\x1b[36m";  // 文字色をシアンに
@@ -122,7 +122,7 @@ namespace etrobocon2023_test {
     // 直進後の走行距離
     int rightCount = Measurer::getRightCount();
     int leftCount = Measurer::getLeftCount();
-    int actual = Mileage::calculateMileage(rightCount, leftCount);
+    double actual = Mileage::calculateMileage(rightCount, leftCount);
 
     EXPECT_EQ(expectedOutput, actualOutput);  // 標準出力でWarningを出している
     EXPECT_EQ(expected, actual);              // 直進前後で走行距離に変化はない
@@ -137,7 +137,7 @@ namespace etrobocon2023_test {
     // 初期値から期待する走行距離を求める
     int initialRightCount = Measurer::getRightCount();
     int initialLeftCount = Measurer::getLeftCount();
-    int expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
+    double expected = Mileage::calculateMileage(initialRightCount, initialLeftCount);
 
     // Warning文
     string expectedOutput = "\x1b[36m";  // 文字色をシアンに
@@ -151,7 +151,7 @@ namespace etrobocon2023_test {
     // 直進後の走行距離
     int rightCount = Measurer::getRightCount();
     int leftCount = Measurer::getLeftCount();
-    int actual = Mileage::calculateMileage(rightCount, leftCount);
+    double actual = Mileage::calculateMileage(rightCount, leftCount);
 
     EXPECT_EQ(expectedOutput, actualOutput);  // 標準出力でWarningを出している
     EXPECT_EQ(expected, actual);              // 直進前後で走行距離に変化はない
